+++ base/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Re-implement functions `next_up` and `next_down`, and expose them through the `utils`` module.
+
 ## 0.4.0
 
 ### Add